requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
amply=0.1.6=pyhd8ed1ab_1
annotated-types=0.7.0=pyhd8ed1ab_1
appdirs=1.4.4=pyhd8ed1ab_1
argparse-dataclass=2.0.0=pyhd8ed1ab_0
attrs=25.1.0=pyh71513ae_0
bc=1.07.1=h7f98852_0
biopython=1.79=py311hd4cff14_3
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py311hfdbb021_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.4=hb9d3cd8_0
ca-certificates=2025.1.31=hbcca054_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py311hf29c0ef_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
click=8.1.8=pyh707e725_0
coin-or-cbc=2.10.12=h8b142ea_1
coin-or-cgl=0.60.7=h516709c_0
coin-or-clp=1.17.8=h1ee7a9c_0
coin-or-osi=0.108.10=haf5fa05_0
coin-or-utils=2.11.11=hee58242_0
coincbc=2.10.12=1_metapackage
colorama=0.4.6=pyhd8ed1ab_1
coloredlogs=15.0.1=pyhd8ed1ab_4
colormath=3.0.0=pyhd8ed1ab_4
conda-inject=1.3.2=pyhd8ed1ab_0
configargparse=1.7=pyhd8ed1ab_1
connection_pool=0.0.3=pyhd3deb0d_0
contourpy=1.3.1=py311hd18a35c_0
cycler=0.12.1=pyhd8ed1ab_1
datrie=0.8.2=py311h9ecbd09_8
docutils=0.21.2=pyhd8ed1ab_1
dpath=2.2.0=pyha770c72_0
expat=2.6.4=h5888daf_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-forge=1=0
fonttools=4.56.0=py311h2dc5d0c_0
freetype=2.12.1=h267a509_2
gfaffix=0.2.0=hc1c3326_2
gfapy=1.2.3=pyhdfd78af_0
gitdb=4.0.12=pyhd8ed1ab_0
gitpython=3.1.44=pyhff2d567_0
graphaligner=1.0.17=hd03093a_0
gsl=2.7=he838d99_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
htslib=1.21=h5efdd21_0
humanfriendly=10.0=pyh707e725_8
humanize=4.12.1=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.2=h59595ed_0
idna=2.10=pyh9f0ad1d_0
immutables=0.21=py311h9ecbd09_0
importlib-metadata=8.6.1=pyha770c72_0
importlib_resources=6.5.2=pyhd8ed1ab_0
intervaltree=3.1.0=pyhd8ed1ab_1
jemalloc=4.5.0=0
jinja2=3.1.5=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_1
jsonschema=4.23.0=pyhd8ed1ab_1
jsonschema-specifications=2024.10.1=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kaleido-core=0.2.1=h3644ca4_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py311hd18a35c_0
krb5=1.21.3=h659f571_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
levenshtein=0.26.1=py311hfdbb021_0
libblas=3.9.0=30_h59b9bed_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=30_he106b2a_openblas
libcurl=8.8.0=hca28451_1
libdeflate=1.20=hd590300_0
libdivsufsort=2.0.2=h7b50bb2_10
libedit=3.1.20250104=pl5321h7949ede_0
libev=4.33=hd590300_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.6=h2dba641_0
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.2=default_he43201b_1000
libiconv=1.18=h4ce23a2_0
libjemalloc=5.3.0=h5888daf_1
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=30_h7ac8fdf_openblas
liblapacke=3.9.0=30_he2f377e_openblas
liblzma=5.6.4=hb9d3cd8_0
liblzma-devel=5.6.4=hb9d3cd8_0
libnghttp2=1.58.0=h47da74e_1
libnsl=2.0.1=hd590300_0
libopenblas=0.3.29=pthreads_h94d23a6_0
libpng=1.6.43=h2797004_0
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtiff=4.6.0=h1dd3fc0_3
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.15=h0b41bf4_0
libxml2=2.12.7=hc051c1a_1
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=19.1.7=h024ca30_0
loguru=0.6.0=py311h38be061_2
make_prg=0.5.0=pyhdfd78af_0
markdown=3.6=pyhd8ed1ab_0
markdown-it-py=3.0.0=pyhd8ed1ab_1
markupsafe=3.0.2=py311h2dc5d0c_1
mathjax=2.7.7=ha770c72_3
matplotlib-base=3.10.0=py311h2b939e6_0
mdurl=0.1.2=pyhd8ed1ab_1
mkl=2024.2.2=ha957f24_16
multiqc=1.27.1=pyhdfd78af_0
munkres=1.0.7=py_1
narwhals=1.27.1=pyhd8ed1ab_0
natsort=8.4.0=pyh29332c3_1
nbformat=5.10.4=pyhd8ed1ab_1
ncurses=6.5=h2d0b736_3
networkx=3.4.2=pyh267e887_2
nspr=4.36=h5888daf_0
nss=3.100=hca3bf56_0
numpy=1.26.4=py311h64a7726_0
odgi=0.9.0=py311hc849397_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.4.1=h7b32b05_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py311h7db5c69_1
patsy=1.0.1=pyhd8ed1ab_1
pggb=0.3.0=hdfd78af_0
pigz=2.8=h2797004_0
pillow=10.3.0=py311h18e6fac_0
pip=25.0.1=pyh8b19718_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
plac=1.4.3=pyhd8ed1ab_1
platformdirs=4.3.6=pyhd8ed1ab_1
plotly=6.0.0=pyhd8ed1ab_0
psutil=6.1.1=py311h9ecbd09_0
pthread-stubs=0.4=hb9d3cd8_1002
pulp=2.8.0=py311h38be061_0
pyaml-env=1.2.2=pyhd8ed1ab_0
pybind11=2.13.6=pyh1ec8472_2
pybind11-global=2.13.6=pyh415d2e4_2
pycparser=2.22=pyh29332c3_1
pydantic=2.10.6=pyh3cfb1c2_0
pydantic-core=2.27.2=py311h9e33e62_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha55dd90_7
python=3.11.0=he550d4f_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pyhd8ed1ab_1
python-fastjsonschema=2.21.1=pyhd8ed1ab_0
python-kaleido=0.2.1=pyhd8ed1ab_0
python-tzdata=2025.1=pyhd8ed1ab_0
python_abi=3.11=5_cp311
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py311h2dc5d0c_2
qhull=2020.2=h434a139_5
rapidfuzz=3.12.1=py311hfdbb021_0
readline=8.2=h8228510_1
referencing=0.36.2=pyh29332c3_0
regex=2024.11.6=py311h9ecbd09_0
requests=2.32.3=pyhd8ed1ab_1
reretry=0.11.8=pyhd8ed1ab_1
rich=13.9.4=pyhd8ed1ab_1
rich-click=1.8.5=pyhd8ed1ab_0
rpds-py=0.22.3=py311h9e33e62_0
samtools=1.21=h50ea8bc_0
scikit-learn=1.3.2=py311hc009520_2
scipy=1.15.2=py311h8f841c2_0
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
seqwish=0.7.11=h43eeafb_0
setuptools=65.6.3=pyhd8ed1ab_0
six=1.17.0=pyhd8ed1ab_0
smart_open=7.1.0=pyhd8ed1ab_0
smmap=5.0.2=pyhd8ed1ab_0
smoothxg=0.6.1=hfb1f815_2
snakemake-interface-common=1.17.4=pyhdfd78af_0
snakemake-interface-executor-plugins=9.3.3=pyhdfd78af_0
snakemake-interface-report-plugins=1.1.0=pyhdfd78af_0
snakemake-interface-storage-plugins=3.3.0=pyhdfd78af_0
snakemake-minimal=8.28.0=pyhdfd78af_0
sortedcontainers=2.4.0=pyhd8ed1ab_1
spectra=0.0.11=pyhd8ed1ab_2
sqlite=3.46.0=h6d4b2fc_0
statsmodels=0.14.4=py311h9f3472d_0
tabulate=0.9.0=pyhd8ed1ab_2
tbb=2021.13.0=hceb3a55_1
threadpoolctl=3.5.0=pyhc1e730c_0
throttler=1.2.2=pyhd8ed1ab_0
tiktoken=0.9.0=py311hf1706b8_0
time=1.9=hb9d3cd8_0
tk=8.6.13=noxft_h4845f30_101
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=pyhd8ed1ab_1
typeguard=4.4.2=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025a=h78e105d_0
unicodedata2=16.0.0=py311h9ecbd09_0
urllib3=2.3.0=pyhd8ed1ab_0
vg=1.63.1=h9ee0642_0
wfmash=0.21.0=h11f254b_0
wheel=0.45.1=pyhd8ed1ab_1
wrapt=1.17.2=py311h9ecbd09_0
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xz=5.6.4=hbcc6ac9_0
xz-gpl-tools=5.6.4=hbcc6ac9_0
xz-tools=5.6.4=hb9d3cd8_0
yaml=0.2.5=h7f98852_2
yte=1.5.7=pyha770c72_0
zipp=3.21.0=pyhd8ed1ab_1
zstandard=0.23.0=py311hbc35293_1
zstd=1.5.6=ha6fb4c9_0
